Added support for additional files to TemplateSignatureRequests
The HelloSign REST API has support for TemplateSignatureRequests with
additional files appended to the end, as documented here:
https://app.hellosign.com/api/reference#send_with_template
https://app.hellosign.com/api/reference#send_embedded_signature_request_with_template

These can be added either as files or file urls. This change allows the
Java SDK to send the file or file_urls with the TemplateSignatureRequests. Per the
API documentation, only one type (file vs. file url) can be specified. This
change does not address enforcing that, as it is not enforced in the SignatureRequest
class that this is modeled on.

Note that this fixes https://github.com/hellosign/hellosign-java-sdk/issues/81